+++ test/dummy/config/environments/test.rb
@@ -4,4 +4,5 @@ Rails.application.configure do
   config.action_controller.perform_caching = false
   config.active_support.deprecation = :log
   config.active_record.maintain_test_schema = false
+  config.active_job.queue_adapter = :inline
 end